Create Skeleton code
Code that we could possibly use as the final code for the robot. I threw in a few things that need to be added and the scratch code that i used for the gps. The compass, rotary's ,infrared sensors, WIFI module, LCD, pin pad & servo motor for unlocking needs to still be added. Most of which should be pretty simple to get working